requirements.txt
fastapi==0.110.2
uvicorn==0.30.1
pandas==2.2.2
torch==2.3.0
torchvision==0.18.0
Pillow==10.4.0
scikit-learn==1.5.1
python-multipart==0.0.9